replace nose with py.test. Fixed DeprecationWarnings (#131)
* replace nose with py.test. Fixed DeprecationWarnings

* no more nose

* test with specific pytest-cov version

* formatting

+++ LFPy/test/test_templatecell.py
@@ -48,7 +48,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -61,7 +61,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -74,7 +74,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -87,7 +87,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -100,7 +100,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -113,7 +113,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -126,7 +126,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
 
@@ -139,7 +139,7 @@ class testTemplateCell(unittest.TestCase):
 
         tvec = stickSimulationTesttvec(**stickParams)
         tvec_numpy = np.linspace(0, stickParams['tstop'],
-                    stickParams['tstop']/stickParams['dt'] + 1)
+                    int(stickParams['tstop']/stickParams['dt']) + 1)
 
         np.testing.assert_equal(tvec, tvec_numpy)
